dom/html/test/file_fullscreen-table.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="UTF-8">
  <title>Test for Bug 1223561</title>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <script type="text/javascript" src="file_fullscreen-utils.js"></script>
</head>
<body>
<table style="background-color: green"></table>
<script>
"use strict";

function ok(condition, msg) {
  opener.ok(condition, "[table] " + msg);
}

function is(a, b, msg) {
  opener.is(a, b, "[table] " + msg);
}

function info(msg) {
  opener.info("[table] " + msg);
}

const gTable = document.querySelector("table");

function begin() {
  info("The default background of window should be white");
  addFullscreenChangeContinuation("enter", enteredFullscreen);
  assertWindowPureColor(window, "white");
  gTable.requestFullscreen();
}

function enteredFullscreen() {
  info("The table with green background should be in fullscreen");
  assertWindowPureColor(window, "green");
  gTable.style = "background: transparent";
  info("When the table becames transparent, the black backdrop should appear");
  assertWindowPureColor(window, "black");
  addFullscreenChangeContinuation("exit", exitedFullscreen);
  document.exitFullscreen();
}

function exitedFullscreen() {
  opener.nextTest();
}
</script>
</body>
</html>
